refactor(rename): difference to array-diff
